refactor key stuff to keys module. remove keycombo class
